backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
catboost==1.2.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
flake8==7.1.0
fonttools==4.53.0
graphviz==0.20.3
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.9.0
mccabe==0.7.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
psycopg2==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
scipy==1.14.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tenacity==8.4.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
